// src/showtime/dto/ipfs.showtime.dto.ts
export class IpfsShowtimeDto {

  name: string;
  artist: string;
  title: string;
  genres: string;
  description: string;
  musicIpfsHash: string;
  playTime: number;
  tier: string;
  no: number;
  imageIpfsHash: string;
  rareImageIpfsHash: string[];
  royaltyPercent: number;
  royaltyAddress: string;
  tokenId: string;
}